% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{BinaryIndexedTree}
\alias{BinaryIndexedTree}
\title{A simple implementation of Binary Indexed Tree as an R6 class.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BinaryIndexedTree
}
\description{
Methods: \itemize{
\item \code{new(init)} Initializes from a vector init.
\item \code{update(location, new.val)} Replaces the item at location with new.val.
\item \code{query(l, r)} Returns the sum of the interval [l, r].
\item \code{show.BIT ()} Returns the Binary Indexed Tree.
\item \code{show.origin()} Returns the vector.
}
}
\examples{
tmp <- BinaryIndexedTree$new(c(2,3,2,5,1))
tmp$update(1,3)
tmp$query(1,5)
tmp$show.origin()
tmp$show.BIT()
}
\keyword{datasets}
